Revert "[TMGP4-85] Create Unit Test"
This reverts commit d1014f5a74f4477e1cfa9c0ce4553667784cccfd.
