+++ userAuth/userAuth/src/main/java/com/apptime/auth/config/securityConfig.java
@@ -1,0 +1,48 @@
+package com.apptime.auth.config;
+
+import org.aspectj.weaver.ast.And;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+import com.apptime.auth.service.AuthUserDetailService;
+
+
+@Configuration
+@EnableWebSecurity
+@EnableGlobalMethodSecurity(prePostEnabled = true)
+public class securityConfig extends WebSecurityConfigurerAdapter {
+	
+	@Autowired
+	AuthUserDetailService userDetailService;
+	protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+		auth.userDetailsService(userDetailService).passwordEncoder(encodePWD());
+	}
+	
+	protected void configure(HttpSecurity http) throws Exception {
+		
+		http.csrf().disable().authorizeRequests()
+        .antMatchers("/signup/**").anonymous()
+        .antMatchers("/dashboard").hasAuthority("USER")
+        .anyRequest().authenticated()
+        .and()
+        .httpBasic();
+		
+     
+      
+	}
+	
+	@Bean
+	public  BCryptPasswordEncoder encodePWD() {
+		return new BCryptPasswordEncoder();
+	}
+
+}
