+++ userAuth/userAuth/src/main/java/com/apptime/auth/controller/UserController.java
@@ -6,9 +6,7 @@ import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
 
-import com.apptime.auth.exception.UserExistingException;
 import com.apptime.auth.model.ResetPasswordRequest;
-import com.apptime.auth.service.SecurityService;
 import com.apptime.auth.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,13 +16,11 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.ui.Model;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.apptime.auth.model.ClientUser;
@@ -37,17 +33,18 @@ import com.apptime.auth.repository.UserRepository;
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class UserController {
 	@Autowired
-	private UserService userService;
+	private UserRepository userRepository;
 
 	@Autowired
 	private BCryptPasswordEncoder passwordEncoder;
 
 	@Autowired
-	private SecurityService securityService;
-
+	private UserService userService;
+	
 	@PostMapping("/signup")
 	public ResponseEntity<ClientUser> signup(@RequestBody Users user) {
-		String password = user.getPassword();
+		
+		String pwd = user.getPassword();
 		if(user.getRoles()== null) {
 			Roles r = new Roles();
 			r.setRole("USER");
@@ -55,13 +52,9 @@ public class UserController {
 			rSet.add(r);
 			user.setRoles(rSet);
 		}
-		try {
-			userService.createUser(user);
-		} catch (UserExistingException e) {
-			return buildErrorResponse(HttpStatus.CONFLICT);
-		}
-		securityService.autoLogin(user.getUsername(), password);
-
+		String encryptPwd = passwordEncoder.encode(pwd);
+		user.setPassword(encryptPwd);
+		userRepository.save(user);
 		return buildSuccessfulResponse(user);
 		
 	}
@@ -69,36 +62,13 @@ public class UserController {
 	//@PreAuthorize("hasAnyRole('ADMIN')")
 	//@CrossOrigin(origins = "*", allowedHeaders = "*")
 	@GetMapping("/dashboard")
-	public  ResponseEntity<ClientUser> dashboard(Principal p) {
-		Users user = userService.findByUsername(p.getName());
+	public  ResponseEntity<ClientUser> login(Principal p) {	
+		Users user = userRepository.findByUsername(p.getName());
 		
-		if (user != null) {
-			return buildSuccessfulResponse(user);
-		} else {
-			return buildErrorResponse(HttpStatus.UNAUTHORIZED);
-		}
-	}
+		//return new ResponseEntity<ClientUser>(new ClientUser(user.getUsername(),user.getEmail()), HttpStatus.OK);
+		
+		return buildSuccessfulResponse(user);
 
-	@PostMapping("/login")
-	public ResponseEntity<ClientUser> login(@RequestBody Users user) {
-		Users userInDb = userService.findByUsername(user.getUsername());
-		if (userInDb != null && passwordEncoder.matches(user.getPassword(), userInDb.getPassword())) {
-			securityService.autoLogin(user.getUsername(), user.getPassword());
-			return buildSuccessfulResponse(userInDb);
-		}
-		return buildErrorResponse(HttpStatus.UNAUTHORIZED);
-	}
-
-	@GetMapping("/login")
-	public String login(Model model, String error, String logout) {
-		System.err.println("logined: " + securityService.findLoggedInUsername());
-		if (error != null)
-			model.addAttribute("error", "Your username and password is invalid.");
-
-		if (logout != null)
-			model.addAttribute("message", "You have been logged out successfully.");
-
-		return "login";
 	}
 
 	@PostMapping("/resetPassword")
@@ -106,7 +76,7 @@ public class UserController {
 		if (StringUtils.isEmpty(request.getUsername()) || StringUtils.isEmpty(request.getOldPassword()) || StringUtils.isEmpty(request.getNewPassword())) {
 			return buildErrorResponse(HttpStatus.BAD_REQUEST);
 		}
-		Users user = userService.findByUsername(request.getUsername());
+		Users user = userRepository.findByUsername(request.getUsername());
 		if (user == null) {
 			// cannot find the user
 			return buildErrorResponse(HttpStatus.NOT_FOUND);
