--- userAuth/userAuth/src/main/java/com/apptime/auth/service/SecurityService.java
@@ -1,7 +1,0 @@
-package com.apptime.auth.service;
-
-public interface SecurityService {
-    String findLoggedInUsername();
-
-    void autoLogin(String username, String password);
-}
